Move deleteOldProjectsFromHTML to logic.js
